Finished 2.2

+++ SharingImages/AppA/app/src/main/java/com/example/appa/ImageGridFragment.java
@@ -1,5 +1,6 @@
 package com.example.appa;
 
+import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
@@ -11,9 +12,14 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.core.content.FileProvider;
+import androidx.core.view.MenuProvider;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Lifecycle;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import java.io.File;
@@ -22,42 +28,69 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.Locale;
 
+
 public class ImageGridFragment extends Fragment
 {
-    private RecyclerView recyclerView;
-    private ImageAdapter imageAdapter;
     private ArrayList<Uri> imageUris = new ArrayList<>();
-    private static final int REQUEST_IMAGE_CAPTURE = 1;
+    private ImageAdapter imageAdapter;
+    private ActivityResultLauncher<Intent> takePictureLauncher;
 
     @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState)
+    {
         View view = inflater.inflate(R.layout.fragment_image_grid, container, false);
-        recyclerView = view.findViewById(R.id.recycler_view);
+        RecyclerView recyclerView = view.findViewById(R.id.recycler_view);
         recyclerView.setLayoutManager(new GridLayoutManager(getContext(), 3));
         imageAdapter = new ImageAdapter(imageUris, this::handleImageClick, this::handleDelete);
         recyclerView.setAdapter(imageAdapter);
         loadImages();
-        setHasOptionsMenu(true);
+
+        takePictureLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(),
+                result ->
+                {
+                    if (result.getResultCode() == android.app.Activity.RESULT_OK)
+                    {
+                        int previousSize = imageUris.size();
+                        imageUris.clear();
+                        loadImages();
+                        int newSize = imageUris.size();
+                        if (newSize > previousSize)
+                        {
+                            imageAdapter.notifyItemRangeInserted(previousSize, newSize - previousSize);
+                        }
+                        else
+                        {
+                            imageAdapter.notifyDataSetChanged();
+                        }
+                    }
+                });
+
+        requireActivity().addMenuProvider(new MenuProvider()
+        {
+            @Override
+            public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater)
+            {
+                menuInflater.inflate(R.menu.menu_main, menu);
+            }
+
+            @Override
+            public boolean onMenuItemSelected(@NonNull MenuItem menuItem)
+            {
+                if (menuItem.getItemId() == R.id.action_take_picture)
+                {
+                    dispatchTakePictureIntent();
+                    return true;
+                }
+                return false;
+            }
+        }, getViewLifecycleOwner(), Lifecycle.State.RESUMED);
+
         return view;
     }
 
-    @Override
-    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
-        inflater.inflate(R.menu.menu_main, menu);
-        super.onCreateOptionsMenu(menu, inflater);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        if (item.getItemId() == R.id.action_take_picture) {
-            dispatchTakePictureIntent();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void dispatchTakePictureIntent() {
+    private void dispatchTakePictureIntent()
+    {
         File photoFile = createImageFile();
         Uri photoUri = FileProvider.getUriForFile(
                 requireContext(),
@@ -66,20 +99,24 @@ public class ImageGridFragment extends Fragment
         );
         Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
         takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoUri);
-        startActivityForResult(takePictureIntent, REQUEST_IMAGE_CAPTURE);
+        takePictureLauncher.launch(takePictureIntent);
     }
 
-    private File createImageFile() {
-        String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss", Locale.US).format(new Date());
-        File storageDir = requireContext().getDir("images", MODE_PRIVATE);
+    private File createImageFile()
+    {
+        String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(new Date());
+        File storageDir = requireContext().getDir("images", Context.MODE_PRIVATE);
         return new File(storageDir, "IMG_" + timeStamp + ".jpg");
     }
 
-    private void loadImages() {
-        File storageDir = requireContext().getDir("images", MODE_PRIVATE);
+    private void loadImages()
+    {
+        File storageDir = requireContext().getDir("images", Context.MODE_PRIVATE);
         File[] files = storageDir.listFiles();
-        if (files != null) {
-            for (File file : files) {
+        if (files != null)
+        {
+            for (File file : files)
+            {
                 Uri uri = FileProvider.getUriForFile(
                         requireContext(),
                         "com.example.appa.fileprovider",
@@ -91,17 +128,20 @@ public class ImageGridFragment extends Fragment
         }
     }
 
-    private void handleImageClick(Uri uri) {
+    private void handleImageClick(Uri uri)
+    {
         String action = requireActivity().getIntent().getAction();
-        if (Intent.ACTION_PICK.equals(action)) {
-            // Return the URI to App B
+
+        if (Intent.ACTION_PICK.equals(action))
+        {
             Intent resultIntent = new Intent();
             resultIntent.setData(uri);
             resultIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-            requireActivity().setResult(RESULT_OK, resultIntent);
+            requireActivity().setResult(android.app.Activity.RESULT_OK, resultIntent);
             requireActivity().finish();
-        } else {
-            // Open the image in the gallery app
+        }
+        else
+        {
             Intent viewIntent = new Intent(Intent.ACTION_VIEW);
             viewIntent.setDataAndType(uri, "image/*");
             viewIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
@@ -109,19 +149,23 @@ public class ImageGridFragment extends Fragment
         }
     }
 
-    private void handleDelete(Uri uri) {
-        File file = new File(uri.getPath());
-        if (file.exists()) {
-            file.delete();
-        }
-    }
+    private void handleDelete(Uri uri)
+    {
+        String path = uri.getPath();
 
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
-            imageUris.clear();
-            loadImages();
+        if (path != null)
+        {
+            File file = new File(path);
+
+            if (file.exists())
+            {
+                boolean deleted = file.delete();
+
+                if (!deleted)
+                {
+                    Log.w("ImageGridFragment", "Failed to delete file: " + uri);
+                }
+            }
         }
     }
 }
